requirements-browser-agent.txt
google-generativeai>=0.8.0
PyYAML>=6.0.1
